api/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-bigquery
python-dotenv
python-jose[cryptography]
